python-services/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
openpyxl
pillow
opencv-python
reportlab
roboflow
pypdf
python-multipart
psycopg2-binary
boto3
requests
geojson
